src/main/java/de/caritas/cob/userservice/api/service/liveevents/LiveEventNotificationService.java
package de.caritas.cob.userservice.api.service.liveevents;

import static de.caritas.cob.userservice.liveservice.generated.web.model.EventType.DIRECTMESSAGE;
import static org.apache.commons.lang3.StringUtils.isNotBlank;

import de.caritas.cob.userservice.liveservice.generated.web.LiveControllerApi;
import java.util.List;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

/**
 * Service class to provide live event triggers to the live service.
 */
@Service
@RequiredArgsConstructor
public class LiveEventNotificationService {

  private final @NonNull LiveControllerApi liveControllerApi;
  private final @NonNull UserIdsProviderFactory userIdsProviderFactory;

  /**
   * Collects all relevant user or consultant ids of chats and sessions and sends a new
   * directmessage to the live service.
   *
   * @param rcGroupId the rocket chat group id used to observe relevant users
   */
  public void sendLiveDirectMessageEventToUsers(String rcGroupId) {
    if (isNotBlank(rcGroupId)) {
      List<String> userIds = this.userIdsProviderFactory.byRocketChatGroup(rcGroupId)
          .collectUserIds(rcGroupId);
      this.liveControllerApi.sendLiveEvent(userIds, DIRECTMESSAGE.toString());
    }
  }

}
